Move project index default fs facade into shared fs utils
The default async fs facade previously lived in src/plugin/src/project-index/fs-facade.js even though it only wrapped fs.promises for the project index API. Hoisting it into src/shared/utils/fs.js keeps the canonical Node-backed facade alongside other filesystem helpers, promotes reuse across packages, and preserves existing plugin exports by re-exporting the shared helper.
